fix: restart_container timeout type isize -> i64
RestartContainerOptions.t expects i64. Using isize was a portability
bug (works on 64-bit, breaks on 32-bit).

+++ src-tauri/src/api/docker.rs
@@ -161,7 +161,7 @@ pub async fn stop_container(
 #[tauri::command]
 pub async fn restart_container(
     id: String,
-    timeout: isize,
+    timeout: i64,
     state: State<'_, AppState>,
 ) -> CmdResult<OkResponse> {
     let ops = ContainerOps::new(&state.docker);
